ESC-925 PR feedback updates
- Consolidated adapter code into adapter/Message.kt
- Rename translations.kt -> Serialization.kt
- Rename jsBridge.kt to Main.kt
